+++ src/main/java/de/darkandblue/neuralnetwork/layer/Dense.java
@@ -36,7 +36,7 @@ public class Dense extends Layer {
   }
   
   @Override
-  public NumpyArray backward(NumpyArray output_gradient, double learning_rate) {
+  public NumpyArray backward(NumpyArray output_gradient, float learning_rate) {
     // TODO: dimension output_gradient = (1, 1)
     NumpyArray weights_gradient = output_gradient.dot(input.transpose());
     NumpyArray input_gradient = weights.transpose().dot(output_gradient);
