+++ src/main/java/de/darkandblue/neuralnetwork/layer/Layer.java
@@ -5,7 +5,7 @@ import de.darkandblue.neuralnetwork.math.NumpyArray;
 public abstract class Layer {
   public abstract NumpyArray forward(NumpyArray input);
   
-  public abstract NumpyArray backward(NumpyArray output_gradient, double learning_rate);
+  public abstract NumpyArray backward(NumpyArray output_gradient, float learning_rate);
   
   public abstract Layer deepCopy();
 }
