+++ src/main/java/de/darkandblue/neuralnetwork/lossfunction/DiceCoefficient.java
@@ -4,13 +4,13 @@ import de.darkandblue.neuralnetwork.math.NumpyArray;
 
 public class DiceCoefficient implements LossFunction {
   @Override
-  public double loss(NumpyArray y_true, NumpyArray y_pred) {
+  public float loss(NumpyArray y_true, NumpyArray y_pred) {
     if (y_true.rows() != y_pred.rows())
       throw new IllegalArgumentException("Rows doesn't match " + y_true.rows() + ", " + y_pred.rows());
     if (y_true.cols() != y_pred.cols())
       throw new IllegalArgumentException("Cols doesn't match " + y_true.cols() + ", " + y_pred.cols());
   
-    double sum = 0;
+    float sum = 0;
     for (int rowIndex = 0; rowIndex < y_true.rows(); rowIndex++) {
       for (int colIndex = 0; colIndex < y_true.cols(); colIndex++) {
         sum += Math.abs(y_pred.data[rowIndex][colIndex] - y_true.data[rowIndex][colIndex]);
@@ -30,7 +30,7 @@ public class DiceCoefficient implements LossFunction {
     
     
     
-    double[][] newData = new double[y_true.rows()][y_pred.cols()];
+    float[][] newData = new float[y_true.rows()][y_pred.cols()];
     for (int rowIndex = 0; rowIndex < y_true.rows(); rowIndex++) {
       for (int colIndex = 0; colIndex < y_true.cols(); colIndex++) {
       }
